ai/Hybrid_Search_Opensource/requirements.txt
azure-search-documents==11.4.0
azure-identity==1.16.1
openai==1.45.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.32
faiss-cpu==1.7.4
rank-bm25==0.2.2
azure-core==1.29.4
azure-ai-formrecognizer==3.3.0
azure-storage-blob==12.19.1
numpy==1.26.4
jupyter==1.0.0  # optional: only if you want notebooks
